requirements.txt
adium-theme-ubuntu==0.3.4
alembic==0.9.9
apt-xapian-index==0.45
astroid==1.0.1
Babel==2.3.4
backports.csv==1.0.5
backports.functools-lru-cache==1.5
beautifulsoup4==4.4.1
blinker==1.3
brotlipy==0.7.0
capstone==3.0.4
cffi==1.11.5
chardet==2.0.1
click==6.6
colorama==0.2.5
command-not-found==0.3
configobj==4.7.2
confusable-homoglyphs==2.0.2
crutch==0.3.0
csscompressor==0.9.5
csvImporter==0.1.3.2
cvxopt==1.1.4
cycler==0.10.0
debtagshw==0.1
decorator==4.2.1
defer==1.0.6
diff-match-patch==20121119
dirspec==13.10
Django==1.11.4
django-admin-export==2.0
django-admin-list-filter-dropdown==1.0.1
django-appconf==1.0.2
django-avatar==4.0.1
django-bootstrap-form==3.4
django-bootstrap3==8.2.3
django-bulk-admin==0.1.1
django-compressor==2.2
django-crispy-forms==1.6.1
django-datetime-widget==0.9.3
django-debug-toolbar==1.5
django-filter==1.0.2
django-graphos==0.3.41
django-import-export==0.5.1
django-initial-avatars==0.9
django-messages==0.5.3
django-registration==2.3
django-registration-redux==1.8
django-report-utils==0.3.15
django-tables2==1.13.0
django-tables2-column-shifter==0.3.1
djangorestframework==3.7.3
dnspython==1.11.1
docutils==0.11
duplicity==0.6.23
enum34==1.1.6
et-xmlfile==1.0.1
extras==0.0.3
fixtures==2.0.0
Flask==0.11.1
Flask-Babel==0.11.1
Flask-Gravatar==0.4.2
Flask-HTMLmin==1.3.2
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
Flask-Paranoid==0.1.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
fluendo-eula==0.6
freetype-py==1.1
fuse-python==0.2.1
GDAL==1.10.1
h5json==1.1.3
h5py==2.7.1
h5pyViewer==0.0.1.6
hdf-compass==0.7.0b1
HDFconvert==1.0.1
html5lib==1.0b3
htmlmin==0.1.12
httplib2==0.8
importlib==1.0.4
ipython==1.2.1
itsdangerous==0.24
jdcal==1.3
Jinja2==2.7.3
joblib==0.7.1
jpeg4py==0.1.4
linecache2==1.0.0
lockfile==0.8
logilab-common==0.61.0
lxml==3.3.3
Mako==1.0.7
MarkupSafe==0.23
matplotlib==2.1.2
mercurial==2.8.2
MySQL-python==1.2.3
networkx==2.1
nibabel==1.2.2
nose==1.3.1
numexpr==2.2.2
numpy==1.8.2
oauthlib==0.6.1
odfpy==1.3.5
oneconf==0.3.7.14.4.1
openpyxl==2.4.8
openshot==1.4.3
PAM==0.4.2
pandas==0.13.1
passlib==1.6.2
patsy==0.2.1
pbr==1.9.1
pep8==1.4.6
pexpect==3.1
pgadmin4==2.1
Pillow==2.3.0
piston-mini-client==0.7.5
pkgconfig==1.2.2
ply==3.11
pp==1.6.4
psutil==1.2.1
PsychoPy==1.79.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycparser==2.18
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pydicom==0.9.7
pyflakes==0.8.1
pygame===1.9.1release
pyglet==1.1.4
Pygments==1.6
pygobject==3.12.0
pygubu==0.9.8.1
PyInstaller==3.2.1
pylint==1.1.0
pymvpa2==2.0.1
pyo==0.6.8
PyOpenGL==3.0.2
pyOpenSSL==0.13
pyparsing==2.2.0
pyrsistent==0.11.13
pyserial==2.6
pysmbc==1.0.14.1
pyspatialite==3.0.1
python-apt===0.9.3.5ubuntu2
python-dateutil==2.5.0
python-debian===0.1.21-nmu2ubuntu2
python-distutils-extra==2.38
python-editor==1.0.3
python-mimeparse==1.5.1
PyTurboJPEG==1.1.2
pytz==2017.2
PyWavelets==0.2.2
pyxdg==0.25
PyYAML==3.12
pyzmq==14.0.1
rcssmin==1.0.6
reportlab==3.0
requests==2.2.1
rjsmin==1.0.12
roman==2.0.0
rope==0.9.2
scikit-learn==0.14.1
scipy==0.13.3
sessioninstaller==0.0.0
simplegeneric==0.8.1
simplejson==3.11.1
six==1.5.2
slimit==0.8.1
software-center-aptd-plugins==0.0.0
speaklater==1.3
Sphinx==1.2.2
spyder==2.2.5
SQLAlchemy==1.0.14
sqlparse==0.2.3
statsmodels==0.5.0
subprocess32==3.2.7
system-service==0.1.6
tables==3.1.1
tablib==0.12.1
testtools==2.3.0
tornado==3.1.1
traceback2==1.4.0
Twisted-Core==13.2.0
Twisted-Web==13.2.0
uncompyle2==1.1
unicodecsv==0.14.1
unittest2==1.1.0
unity-lens-photos==1.0
urllib3==1.7.1
vboxapi==1.0
virtualenv==15.1.0
VTK==5.8.0
Werkzeug==0.9.6
WTForms==2.0.2
xdiagnose===3.6.3build2
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.0.5
